backend/requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.30
uvicorn==0.30.1
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
email-validator==2.1.1
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
asyncio==3.4.3
pytest-mock==3.14.0
pytest-cov==5.0.0
textstat==0.7.1
nltk==3.8.1
scikit-learn==1.5.0
joblib==1.4.2
